refactor(desktop): decompose App.tsx into hooks (#234)
App.tsx was a 722-line god-component and the epicenter of the #210/#212/#220
fresh-ref crash class. Extract its three orchestration concerns into custom
hooks (the existing hooks/ pattern) so each effect's dependency array is
isolated, and pull the bulk-YOLO cascade out of the watcher callback into a
pure, unit-testable function.

- hooks/useTikiFileSync — Rust watcher → state reload + transition toasts +
  GitHub refresh triggers + tikiStateStore sync
- hooks/useGithubFreshness — focus/visibility refresh + optional
  rate-limit-gated poll
- hooks/useBulkYoloCascade — stable callback wrapping the pure cascade
- utils/bulkYoloCascade.advanceBulkYoloOnStateChange — the cascade logic,
  now UNIT-TESTED (4 cases); previously untestable inside the listen callback
- utils/tikiStateSync — TikiState type, detectStateChanges,
  normalizeRecentReleases, and syncTikiStateStore (collapses the 3 duplicated
  inline sync blocks that lived in loadState x2 + the file handler)
- utils/scheduleRefresh — shared per-surface debounce
- freshRefGuard now guards useTikiFileSync + useGithubFreshness

Behavior preserved: every extracted callback/ref is stable, so the effects
re-subscribe only on activeProject / settings change, exactly as before.
App.tsx 722 -> 460 lines.

pnpm build (tsc -b) clean; desktop tests 177 -> 181 (+4 cascade).

Closes #234
