+++ apps/desktop/src/hooks/index.ts
@@ -1,2 +1,5 @@
 export { useCommandActions } from './useCommandActions';
 export { useStaleWorkDetection } from './useStaleWorkDetection';
+export { useTikiFileSync } from './useTikiFileSync';
+export { useGithubFreshness } from './useGithubFreshness';
+export { useBulkYoloCascade } from './useBulkYoloCascade';
